requirements.txt
sentence-transformers
faiss-cpu
pdfplumber
openai
numpy
tqdm
python-dotenv
transformers
sentencepiece
